Prevent cobra-init from overwriting files
Running `cobra-cli init` or `cobra-cli add` would overwrite existing
files without warning, which could result in unexpected loss of data.
This commit modifies cobra-cli so that by default it will not
overwrite files, and adds the `--force` flag to the `init` and `add`
subcommands to allow it to overwrite files.

Closes #59